lint fixes

+++ examples/booking/mintnbuy.go
@@ -174,7 +174,7 @@ func main() {
 			return
 		}
 
-		priceBigInt, err = booking.ConvertPriceToBigInt(price.Value, price.Decimals, int32(tokenDecimals))
+		priceBigInt, err = booking.ConvertPriceToBigInt(price.Value, price.Decimals, tokenDecimals)
 		paymentToken = contractAddress
 	case *typesv2.Currency_IsoCurrency:
 		priceBigInt, err = booking.ConvertPriceToBigInt(price.Value, price.Decimals, booking.ISODecimals)
